+++ packages/headless/src/insight.index.ts
@@ -30,11 +30,26 @@ export type {
 } from './controllers/controller/headless-controller';
 export {buildController} from './controllers/controller/headless-controller';
 
-export type {InsightInterfaceState} from './features/insight-interface/insight-interface-state';
-export type {InsightInterface} from './controllers/insight-interface/insight-interface';
-export {buildInsightInterface} from './controllers/insight-interface/insight-interface';
+export type {
+  NumericFacetBreadcrumb,
+  FacetBreadcrumb,
+  DateFacetBreadcrumb,
+  CategoryFacetBreadcrumb,
+  StaticFilterBreadcrumb,
+  Breadcrumb,
+  BreadcrumbValue,
+  BreadcrumbManagerState,
+  BreadcrumbManager,
+  DeselectableValue,
+} from './controllers/insight/breadcrumb-manager/headless-insight-breadcrumb-manager';
+export {buildBreadcrumbManager} from './controllers/insight/breadcrumb-manager/headless-insight-breadcrumb-manager';
+
 export type {
   InsightSearchBox,
   SearchBoxState,
 } from './controllers/insight/search-box/headless-insight-search-box';
 export {buildInsightSearchBox} from './controllers/insight/search-box/headless-insight-search-box';
+
+export type {InsightInterfaceState} from './features/insight-interface/insight-interface-state';
+export type {InsightInterface} from './controllers/insight-interface/insight-interface';
+export {buildInsightInterface} from './controllers/insight-interface/insight-interface';